Fix build error

+++ src/main/java/net/vulkanmod/vulkan/memory/MemoryTypes.java
@@ -104,12 +104,12 @@ public class MemoryTypes {
 
         @Override
         void copyToBuffer(Buffer buffer, long bufferSize, ByteBuffer byteBuffer) {
-            VUtil.memcpy(byteBuffer, buffer.data.getByteBuffer(0, (int) buffer.bufferSize), (int) bufferSize, buffer.getUsedBytes());
+            VUtil.memcpy(byteBuffer, buffer.data.getByteBuffer(0, buffer.bufferSize), (int) bufferSize, buffer.getUsedBytes());
         }
 
         @Override
         void copyFromBuffer(Buffer buffer, long bufferSize, ByteBuffer byteBuffer) {
-            VUtil.memcpy(buffer.data.getByteBuffer(0, (int) buffer.bufferSize), byteBuffer, 0);
+            VUtil.memcpy(buffer.data.getByteBuffer(0, buffer.bufferSize), byteBuffer, 0);
         }
 
         @Override
@@ -133,7 +133,7 @@ public class MemoryTypes {
         }
 
         void copyToBuffer(Buffer buffer, long dstOffset, long bufferSize, ByteBuffer byteBuffer) {
-            VUtil.memcpy(byteBuffer, buffer.data.getByteBuffer((int) 0, (int) buffer.bufferSize), (int) bufferSize, dstOffset);
+            VUtil.memcpy(byteBuffer, buffer.data.getByteBuffer(0, buffer.bufferSize), (int) bufferSize, dstOffset);
         }
 
         void copyBuffer(Buffer src, Buffer dst) {
